Create README.md
In the case of the oracle WebLogic server, it is mandatory that  it check for web.xml and WEB-INF folder if not available they display a Deployment Exception. In the case of the Tomcat server. If the file is available then the parser will check whether the XML file is available, then the parser will check web.xml is valid or not. If it is not valid display an Error message.

Whenever the client types the URL and clicks on the enter button. Browser takes the URL and converts that into httprequest format and sends the request to the server. Now the server has created 2 objects request and response objects. Now the server has open httprequestformat and read all the contents and store inside the request object.

![](servlet_basics/images/)